Update default-timer-jobs-in-sharepoint-server-2019.md
The comment block at the top of the source code for this timer job states:
    /// <summary>
    /// Search Change Log Generator timer job definition
    /// This timer job is responsible to communicate to the termstore, any changes to the physical page.
    /// We cannot do this in an event receiver because event receivers cannot pick up certain changes like permissions etc.,
    /// Only termstore changes are picked up by the search connector and furls are only crawled by the connector (in most cases)
    /// </summary>
The current description does not properly reflect that the timer job is tied to Friendly Urls (fulrs) in publishing pages. I think the proposed change more adequately shows what the timer job is used for.